# management _sysytem/app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App Name API Documentation",
...
        "version": "1.0.0",
        "x-logo": {
            "url": "http://laravel-project-template.test/logo.png",
            "backgroundColor": "#FFFFFF",
            "altText": "LDE"
        }
    },
    "servers": [
        {
            "url": "http://laravel-project-template.test",
            "description": "Production Server"
        },
        {
            "url": "http://laravel-project-template.test",
            "description": "Staging Server"
        },
        {
            "url": "http://laravel-project-template.test/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/candidacies": {
            "get": {
                "tags": [
                    "Candidacy"
                ],
                "summary": "Get list of candidacies.",
                "description": "Returns list of candidacies.",
                "operationId": "getAllCandidaciesData",
                "parameters": [
                    {
                        "name": "final_status",
                        "in": "path",
                        "description": "final_status of candidacy",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "active"
                        }
                    },
                    {
                        "name": "candidate_name",
                        "in": "path",
                        "description": "name of candidate",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Rajiv Gandhi"
                        }
                    },
                    {
                        "name": "position",
                        "in": "path",
                        "description": "position of role",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Software Engineer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucessfull Retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Candidacies retreived successfully"
                                        },
                                        "data": {
                                            "description": "Array of data of all candidacies",
                                            "properties": {
                                                "candidacies": {
                                                    "description": "get list of all candidacie",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Candidacy"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/paginationMeta"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidacy"
                ],
                "summary": "Creates a new candidacy",
                "description": "Returns key of candidacy",
                "operationId": "createNewCandidacy",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "$ref": "#/components/schemas/Candidate/properties/name"
                                    },
                                    "email": {
                                        "$ref": "#/components/schemas/Candidate/properties/email"
                                    },
                                    "position": {
                                        "$ref": "#/components/schemas/Candidacy/properties/position"
                                    },
                                    "metadata": {
                                        "$ref": "#/components/schemas/Candidate/properties/metadata"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucessfull Creation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Candidacy initiated successfully"
                                        },
                                        "data": {
                                            "description": "data of candidacy",
                                            "properties": {
                                                "candidacy_key": {
                                                    "description": "candidacy key to access it",
                                                    "type": "string",
                                                    "example": "UzMOt5LIEvzkfFSh"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/candidacies/{candidacy}": {
            "get": {
                "tags": [
                    "Candidacy"
                ],
                "summary": "Get data of one candidacy.",
                "description": "Returns one candidacy.",
                "operationId": "getOneCandidacyData",
                "parameters": [
                    {
                        "name": "candidacy",
                        "in": "path",
                        "description": "Candidacy key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucessfull Retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Candidacy retreived successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Candidacy"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/candidacies/{candidacy}/closeStage": {
            "post": {
                "tags": [
                    "StageAssignment"
                ],
                "summary": "Closes stage of candidacy",
                "description": "Returns success response if stage closed",
                "operationId": "closeStageOfCandidacy",
                "parameters": [
                    {
                        "name": "candidacy",
                        "in": "path",
                        "description": "Key of Candidacy",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "stage_name": {
                                        "$ref": "#/components/schemas/CandidacyHistory/properties/stage_name"
                                    },
                                    "stage_closing_reason": {
                                        "description": "API for closing stage",
                                        "type": "string",
                                        "example": "Completed the stage"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Closing Successfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Stage closed successfully."
                                        },
                                        "data": {
                                            "description": "API for closing stage",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/candidacies/{candidacy}": {
            "get": {
                "tags": [
                    "AdminCandidacy"
                ],
                "summary": "Get data of one candidacy.",
                "description": "Returns one candidacy.",
                "operationId": "getOneCandidacyData",
                "parameters": [
                    {
                        "name": "candidacy",
                        "in": "path",
                        "description": "Candidacy key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucessfull Retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Candidacy retreived successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Candidacy"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "response5xx": {
                "description": "Response for 5xx type of errors",
                "properties": {
                    "success": {
                        "description": "Shows status of operation",
                        "type": "boolean",
                        "default": false
                    },
                    "message": {
                        "description": "Error message to be shown to user",
                        "type": "string",
                        "example": "Database error."
                    },
                    "error": {
                        "description": "Full error for debugging",
                        "example": {
                            "key1": "data1",
                            "key2": "data2"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "Header": {
                "type": "apiKey",
                "description": "This API uses TOKEN KEY authentication\n\nKey MUST be provided for each request that requires authentication.",
                "name": "apikey",
                "in": "header",
                "scheme": "https"
            },
            "Token": {
                "type": "apiKey",
                "description": "This API uses **TOKEN KEY** authentication\n\nKey `MUST` be provided for each request that requires authentication.",
                "name": "apikey",
                "in": "query",
                "scheme": "https"
            }
        }
    },
...
}